Refactor and repair tests that are supposed to test the RegistrationController
