+++ tests/Feature/Store/EditPageTest.php
@@ -7,7 +7,6 @@ use App\Centre;
 use App\Child;
 use App\CentreUser;
 use App\Registration;
-use Auth;
 use Carbon\Carbon;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use URL;
@@ -137,7 +136,6 @@ class EditPageTest extends StoreTestCase
         }
     }
 
-
     /** @test */
     public function itShowsALogoutButton()
     {
@@ -189,66 +187,6 @@ class EditPageTest extends StoreTestCase
     }
 
     /** @test */
-    public function itWillRejectLeavingWithoutAReason()
-    {
-        $this->actingAs($this->centreUser, 'store')
-            ->visit(URL::route('store.registration.edit', $this->registration->id))
-            ->press('Remove this family')
-            ->press('Yes')
-            // Still see header of leaving popup
-            ->see('Reason for leaving')
-            // Still see the button - will prove that the family is still in scheme
-            ->see('Remove this family')
-        ;
-    }
-
-    /** @test */
-    public function itWillAcceptLeavingWithAReason()
-    {
-        $this->actingAs($this->centreUser, 'store')
-            ->visit(URL::route('store.registration.edit', $this->registration->id))
-            ->press('Remove this family')
-            ->select(config('arc.leaving_reasons')[0], 'leaving_reason')
-            ->press('Yes')
-            ->seePageIs(route('store.registration.index'))
-        ;
-    }
-
-    /** @test */
-    public function itWillNotAcceptAnInvalidLeavingReason()
-    {
-        $response = $this->actingAs($this->centreUser, 'store')
-            ->visit(URL::route('store.registration.edit', $this->registration->id))
-            ->press('Remove this family')
-            ->call(
-                'PUT',
-                route('store.registration.family', $this->registration->family->id),
-                ['leaving_reason' => 'Not a good one']
-            )
-        ;
-        $this->assertResponseStatus(302);
-        $this->assertEquals('The given data was invalid.', $response->exception->getMessage());
-    }
-
-    /** @test */
-    public function itWillRejectUpdatesIfFamilyHasLeft()
-    {
-        $family = $this->registration->family;
-        $family->leaving_on = Carbon::now();
-        $family->leaving_reason = config('arc.leaving_reasons')[0];
-        $family->save();
-        $this->actingAs($this->centreUser, 'store')
-            ->visit(URL::route('store.registration.edit', $this->registration->id))
-            // Throws Authorization exception 403. Expecting this exception doesn't seem to help.
-            // Not sure this is the desired behaviour and it makes untestable. We need to handle gracefully.
-            //->press('Save Changes')
-        ;
-        //$this->assertResponseStatus(403);
-    }
-
-        /**
-     * @test
-     */
     public function childrensDOBsGiveExpectedAge()
     {
         // Set Carbon::now to 01/01/2018
